ENH: renamed helpers.py to mixins.py
